Uploading bos_stress_index shapefile
The Boston Stress Index shapefile contains geospatial data on a 'stress index' developed  to identify sidewalks in the city of Boston which from the point of view of social distancing, are most or least suited for use. Two parameters decide the stress index: daytime population density and sidewalk width. A negative index indicates higher sidewalk capacity for the incident population, and a positive index indicates a lower sidewalk capacity for the incident population. The data are normalized to be between 0 and 1.